% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{cdl.dbf}
\alias{cdl.dbf}
\title{CDL color scheme}
\format{A data frame with 255 rows and 6 variables.}
\usage{
cdl.dbf
}
\description{
A dataset containing the codebook and color sheme of CDL points.
}
\examples{
dplyr::glimpse(cdl.dbf)
}
\keyword{datasets}
