% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilis.r
\name{cdlpal}
\alias{cdlpal}
\title{CDL color palette}
\usage{
cdlpal(x)
}
\arguments{
\item{x}{a integer value or vector of CDL values}
}
\value{
color code used in CDL for given CDL value(s)
}
\description{
Obtain the color code for given CDL value.
}
\examples{
cdlpal(0:10)
}
