% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCDL.r
\name{GetCDLFile}
\alias{GetCDLFile}
\title{Get CDL raster}
\usage{
GetCDLFile(year, b)
}
\arguments{
\item{year}{one of the available CDL years}

\item{b}{a spatial object in the projection of WGS84}
}
\value{
A CDL raster within the given extent
}
\description{
Get CDL raster file by spatial extent in WGS84.
}
\examples{
## not run
# GetCDLFile(2017, poly)
}
\seealso{
\code{\link{GetCDLValue}}
}
