% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCDL.r
\name{GetCDLValue}
\alias{GetCDLValue}
\title{Get CDL point}
\usage{
GetCDLValue(year, lon, lat)
}
\arguments{
\item{year}{one of the available CDL years}

\item{lon}{longitude in WGS84}

\item{lat}{latitude in WGS84}
}
\value{
A list with the CDL value, category and color
}
\description{
Get CDL point information by longitude and latitude in WGS84.
}
\examples{
## not run
# GetCDLValue(2017, -93.65, 42.03)
}
\seealso{
\code\link{GetCDLFile}
}
