% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCDL.r
\name{GetSDLValue}
\alias{GetSDLValue}
\title{Get SDL point}
\usage{
GetSDLValue(lon, lat)
}
\arguments{
\item{lon}{longitude in WGS84}

\item{lat}{latitude in WGS84}
}
\value{
A list with the soil musym, mukey, muname, muacres
}
\description{
Get soil point information by longitude and latitude in WGS84.
}
\examples{
## not run
# GetSDLValue(-93.65, 42.03)
}
\seealso{
\code\link{GetCDLValue}
}
