% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{poly}
\alias{poly}
\title{An example spatial polygon from soil survey data}
\format{A SpatialPolygonsDataFrame with 7 parital map unit polygons. The data slot has 7 rows and 8 variables.}
\usage{
poly
}
\description{
An example spatial polygon from soil survey data
}
\keyword{datasets}
