% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.r
\name{runTool}
\alias{runTool}
\title{Run the overlay tool}
\usage{
runTool()
}
\description{
Run the overlay tool
}
\examples{
## not run
# runTool()
## print the directory containing the code for the application
system.file("shiny-examples", "overlay", package = "viscover")
}
