% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileinPoly.r
\name{TileinPoly}
\alias{TileinPoly}
\title{Overlay raster tile and polygon}
\usage{
TileinPoly(tile, poly)
}
\arguments{
\item{tile}{a raster object}

\item{poly}{a spatial polygon object in WGS84}
}
\value{
a dataframe with the counts of tile points in the polygon
}
\description{
Obtain the spatial intersection of a rater tile and a spatial polygon.
}
\examples{
library(dplyr)
TileinPoly(tile, poly)
}
